Relocating the abstract homebridge services.
