+++ src/services/pauseSwitch.ts
@@ -2,7 +2,7 @@ import { API, CharacteristicSetCallback, HAPStatus, PlatformAccessory } from 'ho
 import { AutomowerContext } from '../automowerAccessory';
 import { PlatformLogger } from '../diagnostics/platformLogger';
 import { MowerState } from '../model';
-import { AbstractSwitch, Switch } from './abstractSwitch';
+import { AbstractSwitch, Switch } from './homebridge/abstractSwitch';
 import { MowerControlService } from './husqvarna/automower/mowerControlService';
 import { MowerIsPausedPolicy } from './policies/mowerIsPausedPolicy';
