+++ test/services/homebridge/abstractSwitch.spec.ts
@@ -2,9 +2,9 @@ import { Characteristic, CharacteristicEventTypes, CharacteristicSetCallback, Ch
 import { API, PlatformAccessory } from 'homebridge';
 import { It, Mock, Times } from 'moq.ts';
 
-import { AutomowerContext } from '../../src/automowerAccessory';
-import { PlatformLogger } from '../../src/diagnostics/platformLogger';
-import { NameMode } from '../../src/services/abstractSwitch';
+import { AutomowerContext } from '../../../src/automowerAccessory';
+import { PlatformLogger } from '../../../src/diagnostics/platformLogger';
+import { NameMode } from '../../../src/services/homebridge/abstractSwitch';
 import { SwitchSpy } from './switchSpy';
 
 describe('AbstractSwitch', () => {
