+++ test/services/scheduleSwitchImpl.spec.ts
@@ -5,7 +5,7 @@ import { It, Mock, Times } from 'moq.ts';
 import { AutomowerContext } from '../../src/automowerAccessory';
 import { PlatformLogger } from '../../src/diagnostics/platformLogger';
 import { Activity, Calendar, Mode, MowerState, Planner, RestrictedReason, State } from '../../src/model';
-import { NameMode } from '../../src/services/abstractSwitch';
+import { NameMode } from '../../src/services/homebridge/abstractSwitch';
 import { MowerControlService } from '../../src/services/husqvarna/automower/mowerControlService';
 import { ScheduleEnabledPolicy } from '../../src/services/policies/scheduleEnabledPolicy';
 import { ScheduleSwitchImplSpy } from './scheduleSwitchImplSpy';
